requirements.txt
aiogram
python-dotenv
httpx
pytest
fastapi
uvicorn[standard]
pydantic
